fix and feat: now zod aply validation and sonner send notifications

+++ frontend/app/bacteria/deleteBacteria.tsx
@@ -1,23 +1,43 @@
-"use client"
+"use client";
 
-import { useRouter } from "next/navigation"
+import { useRouter } from "next/navigation";
+
+import { toast } from "sonner";
 
 interface DeleteBacteriaProps {
-  id: number
+  id: number;
 }
 
-export default function deleteBacteria({ id }: DeleteBacteriaProps) {
-  const router = useRouter()
+export default function DeleteBacteria({ id }: DeleteBacteriaProps) {
+  const router = useRouter();
 
   async function handleDelete() {
-    try {
-      await fetch(`${process.env.NEXT_PUBLIC_API_URL}/bacteria/${id}`, {
-        method: "DELETE",
-      })
+    const confirmed = confirm("Tem certeza que deseja excluir esta bactéria?");
 
-      router.refresh()
+    if (!confirmed) {
+      return;
+    }
+
+    try {
+      const response = await fetch(
+        `${process.env.NEXT_PUBLIC_API_URL}/bacteria/${id}`,
+        {
+          method: "DELETE",
+        },
+      );
+
+      if (!response.ok) {
+        toast.error("Erro ao excluir bactéria");
+        return;
+      }
+
+      toast.success("Bactéria excluída com sucesso!");
+
+      router.refresh();
     } catch (error) {
-      console.log(error)
+      console.log(error);
+
+      toast.error("Erro interno do servidor");
     }
   }
 
@@ -25,5 +45,5 @@ export default function deleteBacteria({ id }: DeleteBacteriaProps) {
     <button className="button-delete" onClick={handleDelete}>
       EXCLUIR
     </button>
-  )
+  );
 }
